rename regions->_regions, fix RTD
