+++ bioframe/__init__.py
@@ -2,7 +2,7 @@
 from __future__ import division, print_function
 from ._version import __version__
 from .schemas import SCHEMAS
-from .region import *
+from ._region import *
 from .ops import *
 from .formats import *
 from .resources import *
