+++ bioframe/formats.py
@@ -10,7 +10,7 @@ import numpy as np
 import pandas as pd
 
 from ._process import run
-from .region import parse_region
+from ._region import parse_region
 from .arrops import argnatsort
 from .schemas import SCHEMAS, BAM_FIELDS, GAP_FIELDS, UCSC_MRNA_FIELDS
